requirements.txt
boto3
drf_yasg
Django
django-storages
python-decouple==3.1
parsedatetime
jwt
setuptools
user-agents